+++ gerenciador-investimentos/src/main/kotlin/com/vitorbasso/gerenciadorinvestimentos/repository/StockJpaRepository.kt
@@ -1,11 +1,10 @@
 package com.vitorbasso.gerenciadorinvestimentos.repository
 
 import com.vitorbasso.gerenciadorinvestimentos.domain.concrete.Stock
+import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface IStockRepository {
+interface StockJpaRepository : JpaRepository<Stock, String>{
     fun findByTickerStartsWith(ticker: String): List<Stock>
-
-    fun findByTicker(ticker: String): Stock?
 }
