+++ gerenciador-investimentos/src/main/kotlin/com/vitorbasso/gerenciadorinvestimentos/repository/concrete/StockRepository.kt
@@ -1,8 +1,10 @@
-package com.vitorbasso.gerenciadorinvestimentos.repository
+package com.vitorbasso.gerenciadorinvestimentos.repository.concrete
 
 import com.vitorbasso.gerenciadorinvestimentos.domain.concrete.Stock
 import com.vitorbasso.gerenciadorinvestimentos.dto.api.response.Quote
 import com.vitorbasso.gerenciadorinvestimentos.integration.YahooApiIntegration
+import com.vitorbasso.gerenciadorinvestimentos.repository.IStockRepository
+import com.vitorbasso.gerenciadorinvestimentos.repository.StockJpaRepository
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Repository
 import java.math.BigDecimal
@@ -10,20 +12,20 @@ import java.time.LocalDateTime
 
 @Repository
 class StockRepository(
-    private val dbStockRepository: IStockRepository,
+    private val dbStockJpaRepository: StockJpaRepository,
     private val yahooApiIntegration: YahooApiIntegration
-) {
+) : IStockRepository {
 
     companion object {
         const val STOCK_TTL: Long = 5 // in minutes
     }
 
-    fun findByTickerStartsWith(ticker: String): List<Stock> {
+    override fun findByTickerStartsWith(ticker: String): List<Stock> {
         TODO()
     }
 
-    fun findByTicker(ticker: String) = this.dbStockRepository.findByIdOrNull(ticker).let {
-        if (isStockInvalid(it)) this.dbStockRepository.save(this.yahooApiIntegration.getQuote(ticker).getEntity())
+    override fun findByTicker(ticker: String) = this.dbStockJpaRepository.findByIdOrNull(ticker).let {
+        if (isStockInvalid(it)) this.dbStockJpaRepository.save(this.yahooApiIntegration.getQuote(ticker).getEntity())
         else it
     }
 
@@ -33,7 +35,7 @@ class StockRepository(
 }
 
 private fun Quote.getEntity() = Stock(
-    ticker = this.symbol,
+    ticker = this.symbol.toUpperCase(),
     currentValue = BigDecimal.valueOf(this.regularMarketPrice),
     openingValue = BigDecimal.valueOf(this.regularMarketOpen),
     closingValue = BigDecimal.valueOf(this.regularMarketPreviousClose),
