+++ gerenciador-investimentos/src/main/kotlin/com/vitorbasso/gerenciadorinvestimentos/service/proxy/StockServiceProxy.kt
@@ -1,7 +1,6 @@
 package com.vitorbasso.gerenciadorinvestimentos.service.proxy
 
 import com.vitorbasso.gerenciadorinvestimentos.domain.concrete.Stock
-import com.vitorbasso.gerenciadorinvestimentos.dto.response.StockDto
 import com.vitorbasso.gerenciadorinvestimentos.dto.response.StockListDto
 import com.vitorbasso.gerenciadorinvestimentos.service.IStockService
 import com.vitorbasso.gerenciadorinvestimentos.service.concrete.StockService
@@ -12,21 +11,11 @@ class StockServiceProxy(
         private val stockService: StockService
 ): IStockService {
 
-    override fun getStock(ticker: String) = this.stockService.getStock(ticker).getDto()
+    override fun getStock(ticker: String) = this.stockService.getStock(ticker.toUpperCase())
 
     override fun getStockStartingWith(ticker: String)
-            = getListDto(this.stockService.getStockStartingWith(ticker))
+            = getListDto(this.stockService.getStockStartingWith(ticker.toUpperCase()))
 
-    private fun getListDto(tickers: List<Stock>) = StockListDto(tickers.map { it.ticker })
+    private fun getListDto(tickers: List<Stock>) = StockListDto(tickers.map { it.ticker.toUpperCase() })
 
 }
-
-private fun Stock.getDto() = StockDto(
-    ticker = this.ticker,
-    currentValue = this.currentValue,
-    openingValue = this.openingValue,
-    closingValue = this.closingValue,
-    highestValue = this.highestValue,
-    lowestValue = this.lowestValue,
-    variation = this.variation
-)
